Moved CRCLib to its own library
