+++ libraries/CMakeLists.txt
@@ -2,3 +2,4 @@ add_subdirectory(platformlib)
 add_subdirectory(thirdparty)
 add_subdirectory(builddefs)
 add_subdirectory(commonutils)
+add_subdirectory(crclib)
