+++ xash3d_engine/engine/common/common.h
@@ -106,7 +106,6 @@ typedef enum
 #include "CommonUtils/crtlib.h"
 #include "cvar.h"
 #include "con_nprint.h"
-#include "CommonUtils/crclib.h"
 #include "ref_api.h"
 #include "fscallback.h"
