+++ xash3d_engine/engine/common/imagelib/img_png.c
@@ -19,6 +19,7 @@ GNU General Public License for more details.
 #include "CommonUtils/xash3d_mathlib.h"
 #include "img_png.h"
 #include "PlatformLib/Net.h"
+#include "CRCLib/crclib.h"
 
 #if !XASH_WIN32
 #include <netinet/in.h>
