+++ xash3d_engine/engine/server/sv_game.c
@@ -24,6 +24,7 @@ GNU General Public License for more details.
 #include "render_api.h"  // modelstate_t
 #include "ref_common.h"  // decals
 #include <limits.h>
+#include "CRCLib/crclib.h"
 
 #define ENTVARS_COUNT ARRAYSIZE(gEntvarsDescription)
