+++ xash3d_engine/engine/server/sv_init.c
@@ -20,6 +20,7 @@ GNU General Public License for more details.
 #include "voice.h"
 #include "pm_local.h"
 #include "sequence.h"
+#include "CRCLib/crclib.h"
 
 int SV_UPDATE_BACKUP = SINGLEPLAYER_BACKUP;
