+++ xash3d_engine/engine/sprite.h
@@ -18,7 +18,6 @@
 
 #include "BuildDefs/build.h"
 #include <stdint.h>
-#include "synctype.h"
 
 /*
 ==============================================================================
