+++ xash3d_engine/filesystem/CMakeLists.txt
@@ -18,6 +18,7 @@ target_link_libraries(${TARGETNAME_LIB_FILESYSTEM_STDIO} PUBLIC
 	${TARGETNAME_INCLUDES_SDK}
 	${TARGETNAME_LIB_COMMONUTILS}
 	${TARGETNAME_LIB_BUILDDEFS}
+	${TARGETNAME_LIB_CRCLIB}
 )
 
 set_common_library_compiler_settings(${TARGETNAME_LIB_FILESYSTEM_STDIO})
